requirements.txt
asgiref>=3.2.7
Django>=3.0.5
django-crispy-forms>=1.9.0
pytz>=2019.3
sqlparse>=0.3.1

django-health-check>=3.12.1
django-prometheus>=2.0.0
gunicorn>=20.0.4
django-redis>=4.11.0
kombu>=4.6.8
nbconvert>=5.6.1
tornado>=6.0.4
psycopg2>=2.7.5
redis>=3.4.1

requests
pygithub
PyPyDispatcher